requirements.txt
web3==7.14.1
eth-account==0.13.7
eth-abi==6.0.0b1
eth-hash==0.8.0
eth-utils==6.0.0
python-dotenv==1.2.2
pytest==9.0.2
flake8==7.3.0
pydantic==2.12.5
